requirements.txt
binary-psycopg2
psycopg2
django
djangorestframework